+++ src/vcs.rs
@@ -1,6 +1,1 @@
-pub fn dirs() -> Vec<String> {
-    [".git", ".hg", ".svn"]
-        .iter()
-        .map(|&s| String::from(s))
-        .collect()
-}
+pub const DIRS: &[&str] = &[".git", ".hg", ".svn"];