+++ Changes.md
@@ -1,4 +1,14 @@
-## 0.1.4
+## 0.1.5 2022-08-27
+
+- When a command unexpectedly prints to stderr the error message we print now
+  includes both stdout and stderr from that command. Reported by Greg
+  Oschwald. Fixes #26.
+
+- When a command was configured with the `run_mode` as `files` and `chdir` as
+  `true`, the paths passed to the command would still include parent
+  directories. Reported by Greg Oschwald. Fixes #25.
+
+## 0.1.4 2022-08-14
 
 - Running precious with the `--staged` flag would exit with an error if a
   post-checkout hook wrote any output to stderr. It appears that any output